requirements.txt
torch
opencv-python
transformers
Pillow
argparse
os
